test: move dataset for testing elasticindexer to conftest
